file_storage/requirements.txt
Django==2.1.1
django-bootstrap3==10.0.1
django-import-export==1.0.1
celery==4.2.1